requirements.txt
ddddocr
pyautogui
numpy
opencv-python
Pillow